Watame
Her name is a pun on "Wata-Ame" 「綿飴」 which translates to "cotton candy" in Japanese. The beginning of Watame's first name, "Wata," translates to "cotton" 「綿」 in Japanese; cotton having a look and texture very similar to that of wool grown by and sheared from sheep. Her surname, Tsunomaki, comes from Tsuno 「角」which translates to "horn" and Maki 「巻」which can translate to "coil," related to her coiled sheep horns.